Move Mermaid DAG code into its own file
